+++ Elements/src/Geometry/CsgExtensions.cs
@@ -22,20 +22,24 @@ namespace Elements.Geometry
         /// Triangulate this csg and pack the triangulated data into buffers
         /// appropriate for use with gltf.
         /// </summary>
-        internal static GraphicsBuffers Tessellate(this Csg.Solid csg, bool mergeVertices = false)
+        internal static GraphicsBuffers Tessellate(this Csg.Solid csg,
+                                                   bool mergeVertices = false,
+                                                   Func<(Vector3, Vector3, UV, Color), (Vector3, Vector3, UV, Color)> modifyVertexAttributes = null)
         {
-            return Tessellate(new[] { csg }, mergeVertices);
+            return Tessellate(new[] { csg }, mergeVertices, modifyVertexAttributes);
         }
 
         /// <summary>
         /// Triangulate a collection of CSGs and pack the triangulated data into
         /// buffers appropriate for use with gltf. 
         /// </summary>
-        internal static GraphicsBuffers Tessellate(this Csg.Solid[] csgs, bool mergeVertices = false)
+        internal static GraphicsBuffers Tessellate(this Csg.Solid[] csgs,
+                                                   bool mergeVertices = false,
+                                                   Func<(Vector3, Vector3, UV, Color), (Vector3, Vector3, UV, Color)> modifyVertexAttributes = null)
         {
             var buffers = new GraphicsBuffers();
 
-            Tessellate(csgs, buffers, mergeVertices);
+            Tessellate(csgs, buffers, mergeVertices, modifyVertexAttributes);
             return buffers;
         }
 
@@ -43,11 +47,14 @@ namespace Elements.Geometry
         /// Triangulate a collection of CSGs and pack the triangulated data into
         /// a supplied buffers object. 
         /// </summary>
-        internal static void Tessellate(Csg.Solid[] csgs, IGraphicsBuffers buffers, bool mergeVertices = false)
+        internal static void Tessellate(Csg.Solid[] csgs,
+                                        IGraphicsBuffers buffers,
+                                        bool mergeVertices = false,
+                                        Func<(Vector3, Vector3, UV, Color), (Vector3, Vector3, UV, Color)> modifyVertexAttributes = null)
         {
             (Vector3 U, Vector3 V) basis;
 
-            var allVertices = new List<(Vector3, Vector3, UV)>();
+            var allVertices = new List<(Vector3 position, Vector3 normal, UV uv, Color color)>();
             foreach (var csg in csgs)
             {
                 foreach (var p in csg.Polygons)
@@ -69,9 +76,9 @@ namespace Elements.Geometry
                             var v = p.Vertices[i];
                             var uu = basis.U.Dot(v.Pos.X, v.Pos.Y, v.Pos.Z);
                             var vv = basis.V.Dot(v.Pos.X, v.Pos.Y, v.Pos.Z);
-                            vertexIndices[i] = (ushort)GetOrCreateVertex((v.Pos.X, v.Pos.Y, v.Pos.Z),
-                                                                         (n.X, n.Y, n.Z),
-                                                                         (uu, vv),
+                            vertexIndices[i] = (ushort)GetOrCreateVertex(new Vector3(v.Pos.X, v.Pos.Y, v.Pos.Z),
+                                                                         new Vector3(n.X, n.Y, n.Z),
+                                                                         new UV(uu, vv),
                                                                          allVertices,
                                                                          mergeVertices);
                         }
@@ -108,9 +115,9 @@ namespace Elements.Geometry
                             var uu = basis.U.Dot(v.Position.X, v.Position.Y, v.Position.Z);
                             var vv = basis.V.Dot(v.Position.X, v.Position.Y, v.Position.Z);
 
-                            vertexIndices[i] = (ushort)GetOrCreateVertex((v.Position.X, v.Position.Y, v.Position.Z),
-                                                                         (n.X, n.Y, n.Z),
-                                                                         (uu, vv),
+                            vertexIndices[i] = (ushort)GetOrCreateVertex(new Vector3(v.Position.X, v.Position.Y, v.Position.Z),
+                                                                         new Vector3(n.X, n.Y, n.Z),
+                                                                         new UV(uu, vv),
                                                                          allVertices,
                                                                          mergeVertices);
 
@@ -127,7 +134,15 @@ namespace Elements.Geometry
 
             foreach (var v in allVertices)
             {
-                buffers.AddVertex(v.Item1, v.Item2, v.Item3);
+                if (modifyVertexAttributes != null)
+                {
+                    var mod = modifyVertexAttributes(v);
+                    buffers.AddVertex(mod.Item1, mod.Item2, mod.Item3, mod.Item4);
+                }
+                else
+                {
+                    buffers.AddVertex(v.Item1, v.Item2, v.Item3);
+                }
             }
         }
 
@@ -139,20 +154,17 @@ namespace Elements.Geometry
             return basis;
         }
 
-        private static int GetOrCreateVertex((double x, double y, double z) position,
-                                             (double x, double y, double z) normal,
-                                             (double u, double v) uv,
-                                             List<(Vector3 position, Vector3 normal, UV uv)> pts,
+        private static int GetOrCreateVertex(Vector3 position,
+                                             Vector3 normal,
+                                             UV uv,
+                                             List<(Vector3 position, Vector3 normal, UV uv, Color color)> pts,
                                              bool mergeVertices)
         {
-            var pt = new Vector3(position.x, position.y, position.z);
-            var n = new Vector3(normal.x, normal.y, normal.z);
-
             if (mergeVertices)
             {
                 var index = pts.FindIndex(p =>
                 {
-                    return p.position.IsAlmostEqualTo(pt) && p.normal.AngleTo(n) < 45.0;
+                    return p.position.IsAlmostEqualTo(position) && p.normal.AngleTo(normal) < 45.0;
                 });
                 if (index != -1)
                 {
@@ -160,7 +172,7 @@ namespace Elements.Geometry
                 }
             }
 
-            pts.Add((pt, n, uv));
+            pts.Add((position, normal, uv, default(Color)));
             return pts.Count - 1;
         }